// src/models/WorkSkills.ts
type WorkSkills = {
  farming?: number;
  kindling?: number;
  handiwork?: number;
  mining?: number;
  transporting?: number;
  cooling?: number;
  planting?: number;
  lumbering?: number;
  gathering?: number;
  watering?: number;
  medicineProduction?: number;
  generatingElectricity?: number;
};

export default WorkSkills;